Convert join fields to inline columns
This seems to be quite a bit faster locally.

This patch changes a few fields to become different types of columns
rather than joined in tables:

- Metadata.CustomFields has become a jsonb column
- Metadata.Visible has become a boolean column
- Metadata.GrepCodes has become a jsonb column
- Translations has become a jsonb column
- CachedPath has become two columns (One `text[]` for primary paths and a `text[]` for others)

There is the caveat that it uses some postgresql specific features,
which are not supported by h2. The next patch will fix this by making
our tests use a real postgresql instance.