Calculate average grade partially
Instead of fetching all children when updating a grade, lets calculate
it by removing the old grade from the average and adding the new one.

This is _much_ faster, at the cost of being slightly less robust.

Updating a deeply nested test-node went from ~20 seconds to ~80ms on my laptop.